requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.41
psycopg2==2.9.3
alembic==1.7.7
pydantic==1.10.13
solana==0.25.0
base58==2.1.0
bcrypt==3.2.2
pyjwt==2.4.0
slowapi==0.1.0
redis==3.5.3
structlog==21.5.0
jinja2==3.0.3
python-dotenv==0.19.2
aiohttp==3.8.5
aiofiles==0.8.0
python-telegram-bot==13.7
requests==2.28.1
pytest==6.2.5
pytz==2025.2
itsdangerous==2.1.2
celery==5.2.7